# docs/openapi.json
{
...
  "components": {
    "schemas": {
...
      "ErrorEnvelope": {
        "properties": {
          "code": {
            "description": "Stable SCREAMING_SNAKE_CASE machine string. Once shipped, never reused.",
            "type": "string"
          },
          "details": {
            "description": "Optional structured detail. Omitted on auth failures per INV-A4.",
            "nullable": true
          },
          "message": {
            "description": "Human-readable, lowercase, no trailing period.",
            "type": "string"
          },
          "request_id": {
            "description": "UUIDv7 request correlator echoed in the `X-Request-Id` header.",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "type": "object"
      },
...
      "ProjectDto": {
        "properties": {
          "created_at": {
            "description": "RFC 3339 / ISO 8601 UTC timestamp (DSD-622).",
            "type": "string"
          },
          "name": {
            "description": "Operator-supplied display name. Free-form, non-empty.",
            "type": "string"
          },
          "project_id": {
            "description": "Stable UUIDv7 (DSD-623) assigned at registration. Sort-friendly.",
            "type": "string"
          },
          "root": {
            "description": "Canonical absolute path of the project root (ISC-008).",
            "type": "string"
          },
          "status": {
            "description": "Lifecycle: `registered` | `active` | `archived`.",
            "type": "string"
          }
        },
        "required": [
          "project_id",
          "name",
          "root",
          "status",
          "created_at"
        ],
        "type": "object"
      },
      "ProjectListDto": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProjectDto"
            },
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
...
      "RegisterProjectRequest": {
        "properties": {
          "name": {
            "description": "Display name for the project. Trimmed; must be non-empty.",
            "type": "string"
          },
          "root": {
            "description": "Absolute filesystem path to the project workspace.",
            "type": "string"
          }
        },
        "required": [
          "root",
          "name"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Mutagen workflow service. Spec is generated from handler annotations via utoipa per ADR-0002.",
    "license": {
      "name": ""
    },
    "title": "mutagen-service",
    "version": "0.3.3"
  },
  "openapi": "3.0.3",
  "paths": {
    "/projects": {
      "get": {
        "operationId": "list_projects",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectListDto"
                }
              }
            },
            "description": "Registered projects (possibly empty)"
          }
        },
        "tags": [
          "projects"
        ]
      },
      "post": {
        "operationId": "register_project",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDto"
                }
              }
            },
            "description": "Project registered"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Root already registered"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation failure"
          }
        },
        "tags": [
          "projects"
        ]
      }
    },
    "/projects/{project_id}": {
      "delete": {
        "operationId": "archive_project",
        "parameters": [
          {
            "description": "Project UUIDv7",
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Project archived (workspace files left intact)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unknown project_id"
          }
        },
        "tags": [
          "projects"
        ]
      },
      "get": {
        "operationId": "get_project",
        "parameters": [
          {
            "description": "Project UUIDv7",
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDto"
                }
              }
            },
            "description": "Project detail"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unknown project_id"
          }
        },
        "tags": [
          "projects"
        ]
      }
    },
...
  }
}